Introduce SessionPathsFactory.
